+++ packages/pt-webgpu/src/adjointPass.ts
@@ -244,6 +244,7 @@ export class AdjointPass {
         { binding: 16, resource: sb.materialTextureView },
         { binding: 17, resource: sb.materialTextureSampler },
         { binding: 18, resource: { buffer: sb.colorsBuffer } },
+        { binding: 19, resource: sb.materialLinearTextureView },
       ],
     });